pyproject.toml
[tool.poetry]
name = "create-billing"
version = "0.1.0"
description = ""
authors = ["Igor Gerasimov <gerasimov.igor.98@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fast-bitrix24 = "^1.7.1"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
